requirements.txt
mcp>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
chromadb>=0.4.0
openai>=1.0.0
sentence-transformers
numpy>=1.24.0 # we did pip install "numpy<2" for pytorch to sync
asyncio-mqtt>=0.11.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0